requirements.txt
boto3==1.40.64
python-dotenv==1.0.1
mysql-connector-python==9.5.0
pathlib==1.0.1
paramiko==2.12.2
sshtunnel==0.4.0